最后一发

+++ server-commons/src/main/java/com/aionstar/commons/network/packet/ServerPacket.java
@@ -17,7 +17,7 @@ public abstract class ServerPacket {
         this.opcode = opcode;
     }
 
-    public byte getOpcode() {
+    public int getOpcode() {
         return opcode;
     }
